fix cache and user parser

+++ lib/services/user_data.dart
@@ -4,7 +4,7 @@ class UserData {
   final String username;
 
   UserData.fromJson(Map<String, dynamic> json)
-      : proDays = json["subscription"]["days"],
+      : proDays = json["subscription"]["days"].toDouble(),
         avatar = json["profile"]["avatar"] == "http://www.gravatar.com/avatar/"
             ? null
             : json["profile"]["avatar"],
